parser_scan/requirements.txt
selenium
easyocr
pandas
pillow
opencv-python
openpyxl
